+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 - Extend completion and getting docs for symbol-strings with leading literals.
+- Find private vars when using var quote literal.
 
 ### 0.3.15 (2023-06-22)